Removed the need to specify the simulation/read cosmology.
The cosmology (hubble parameter and box size) is now read explicitly
from the SAGE parameter file.

If running in binary, then the user must specify the number of output
files that SAGE wrote to (for a single snapshot).  This is because we
otherwise have no knowledge about how much volume the model spans.